Move extensions to package extension
