+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/extension/Player.kt
@@ -4,6 +4,7 @@
  */
 package ch.srgssr.pillarbox.player.extension
 
+import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
 
 /**
@@ -24,4 +25,37 @@ fun Player.startPlayback() {
         }
     }
     play()
+}
+
+/**
+ * Get a snapshot of the current media items
+ */
+fun Player.getCurrentMediaItems(): List<MediaItem> {
+    if (mediaItemCount == 0) {
+        return emptyList()
+    }
+    val count = mediaItemCount
+    return ArrayList<MediaItem>(count).apply {
+        for (i in 0 until count) {
+            add(getMediaItemAt(i))
+        }
+    }
+}
+
+/**
+ * Get playback speed
+ *
+ * @return [Player.getPlaybackParameters] speed
+ */
+fun Player.getPlaybackSpeed(): Float {
+    return playbackParameters.speed
+}
+
+/**
+ * Current position percent
+ *
+ * @return the current position in percent [0,1].
+ */
+fun Player.currentPositionPercentage(): Float {
+    return currentPosition / duration.coerceAtLeast(1).toFloat()
 }
