+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/extension/PlayerCommands.kt
@@ -4,7 +4,7 @@
  */
 @file:Suppress("unused")
 
-package ch.srgssr.pillarbox.player
+package ch.srgssr.pillarbox.player.extension
 
 import androidx.media3.common.Player
