+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/extension/VideoSize.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2023. SRG SSR. All rights reserved.
  * License information is available from the LICENSE file.
  */
-package ch.srgssr.pillarbox.player
+package ch.srgssr.pillarbox.player.extension
 
 import android.util.Rational
 import androidx.media3.common.VideoSize
